LayoutTests/fast/css/cssText-shorthand-expected.txt

This test ensures WebKit uses shorthand notations for cssText; most of what was originally in this test is now covered by Web Platform Tests

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS normalizeCssText(element.style.cssText) is "border-spacing: 1px 2px"
PASS normalizeCssText(element.style.cssText) is "font-family: sans-serif; font-size: 3em; font-style: italic; font-weight: bold; line-height: 2em"
PASS successfullyParsed is true

TEST COMPLETE
